Generate random censored data

Generate random censored data

generate_censored_data(N, rho, L_prob, direction = c(-1), adj = FALSE,
  adj_sigma = NULL)

Arguments

N
number of rows

rho
true correlation

L_prob
percentage of data to censor

direction
types of censoring to include; acceptable values are -1 (left), 1 (right), 2 (interval) or any combination

adj
whether to adjust values by a normally distributed variable

adj_sigma
if adj is true, this specifies the standard deviation of the adjustment variable

Value

data frame

Details

When multiple types of censoring are requested, the function evenly splits the censored values across all the types. Interval censoring is always done around the center of the distribution.

Examples

# Generate 100 x and y values with a 0.5 correlation and 50% left-censored
generate_censored_data(100, rho = 0.5, L_prob = 0.5, direction = -1)

#>          x_true          x    x_upper x_cens       y_true          y    y_upper
#> 1   -1.01886646 0.02272554 0.02272554     -1 -1.406080040 0.02272554 0.02272554
#> 2    0.61382738 0.61382738 0.61382738      0 -0.171605272 0.02272554 0.02272554
#> 3   -1.58236377 0.02272554 0.02272554     -1 -2.639100637 0.02272554 0.02272554
#> 4    0.39294584 0.39294584 0.39294584      0 -0.402595591 0.02272554 0.02272554
#> 5    1.41641816 1.41641816 1.41641816      0 -0.339857268 0.02272554 0.02272554
#> 6    1.33982257 1.33982257 1.33982257      0  0.649284676 0.64928468 0.64928468
#> 7   -1.29846938 0.02272554 0.02272554     -1 -1.857011373 0.02272554 0.02272554
#> 8    0.05414167 0.05414167 0.05414167      0 -0.482521658 0.02272554 0.02272554
#> 9   -0.32504663 0.02272554 0.02272554     -1 -0.097919497 0.02272554 0.02272554
#> 10  -0.73405756 0.02272554 0.02272554     -1  0.244397360 0.24439736 0.24439736
#> 11  -0.37507641 0.02272554 0.02272554     -1 -0.583959058 0.02272554 0.02272554
#> 12   1.24441966 1.24441966 1.24441966      0 -0.154990803 0.02272554 0.02272554
#> 13   1.65909537 1.65909537 1.65909537      0  1.917632663 1.91763266 1.91763266
#> 14  -1.19157853 0.02272554 0.02272554     -1 -1.633377982 0.02272554 0.02272554
#> 15   0.15947483 0.15947483 0.15947483      0  0.728074687 0.72807469 0.72807469
#> 16  -2.67684051 0.02272554 0.02272554     -1 -0.549990050 0.02272554 0.02272554
#> 17  -0.66450532 0.02272554 0.02272554     -1 -0.239646878 0.02272554 0.02272554
#> 18   0.79658618 0.79658618 0.79658618      0 -0.887695356 0.02272554 0.02272554
#> 19   0.34554773 0.34554773 0.34554773      0  0.594949519 0.59494952 0.59494952
#> 20  -1.32803115 0.02272554 0.02272554     -1 -0.255192895 0.02272554 0.02272554
#> 21  -0.61425101 0.02272554 0.02272554     -1  1.425118252 1.42511825 1.42511825
#> 22   0.08959812 0.08959812 0.08959812      0  0.539051897 0.53905190 0.53905190
#> 23  -1.82567487 0.02272554 0.02272554     -1 -0.433860828 0.02272554 0.02272554
#> 24   0.42564976 0.42564976 0.42564976      0  0.852216310 0.85221631 0.85221631
#> 25   1.58170125 1.58170125 1.58170125      0  1.689285240 1.68928524 1.68928524
#> 26  -0.09553730 0.02272554 0.02272554     -1 -0.073242569 0.02272554 0.02272554
#> 27  -1.11227309 0.02272554 0.02272554     -1 -0.508662077 0.02272554 0.02272554
#> 28   0.11751191 0.11751191 0.11751191      0 -0.145138661 0.02272554 0.02272554
#> 29  -0.45188820 0.02272554 0.02272554     -1 -0.980152279 0.02272554 0.02272554
#> 30  -1.40585123 0.02272554 0.02272554     -1 -1.213701808 0.02272554 0.02272554
#> 31   1.38314812 1.38314812 1.38314812      0  0.236948450 0.23694845 0.23694845
#> 32  -0.27024871 0.02272554 0.02272554     -1  0.575935953 0.57593595 0.57593595
#> 33   0.17017799 0.17017799 0.17017799      0  0.251897618 0.25189762 0.25189762
#> 34   2.05859308 2.05859308 2.05859308      0  0.753476072 0.75347607 0.75347607
#> 35   0.56948386 0.56948386 0.56948386      0 -0.375428204 0.02272554 0.02272554
#> 36  -0.34321561 0.02272554 0.02272554     -1  0.111125980 0.11112598 0.11112598
#> 37  -1.22658302 0.02272554 0.02272554     -1 -2.081785521 0.02272554 0.02272554
#> 38  -0.09837894 0.02272554 0.02272554     -1 -0.385274154 0.02272554 0.02272554
#> 39  -0.71893299 0.02272554 0.02272554     -1  0.176028633 0.17602863 0.17602863
#> 40   0.89066352 0.89066352 0.89066352      0  0.957967957 0.95796796 0.95796796
#> 41   0.14199013 0.14199013 0.14199013      0 -0.020686209 0.02272554 0.02272554
#> 42  -0.13984195 0.02272554 0.02272554     -1 -0.967152119 0.02272554 0.02272554
#> 43  -0.98152368 0.02272554 0.02272554     -1  0.894981939 0.89498194 0.89498194
#> 44  -0.60101115 0.02272554 0.02272554     -1  0.165429049 0.16542905 0.16542905
#> 45  -0.10477275 0.02272554 0.02272554     -1  0.875183950 0.87518395 0.87518395
#> 46   1.72537112 1.72537112 1.72537112      0  3.047152114 3.04715211 3.04715211
#> 47   0.60015277 0.60015277 0.60015277      0 -0.519558057 0.02272554 0.02272554
#> 48   0.24301082 0.24301082 0.24301082      0  0.757610639 0.75761064 0.75761064
#> 49   0.85690968 0.85690968 0.85690968      0 -0.652190154 0.02272554 0.02272554
#> 50  -2.42196925 0.02272554 0.02272554     -1 -0.889227770 0.02272554 0.02272554
#> 51   0.53201511 0.53201511 0.53201511      0  0.961162479 0.96116248 0.96116248
#> 52  -0.27170090 0.02272554 0.02272554     -1 -0.149597488 0.02272554 0.02272554
#> 53   0.39052945 0.39052945 0.39052945      0 -0.747482946 0.02272554 0.02272554
#> 54   0.29561585 0.29561585 0.29561585      0 -0.262399283 0.02272554 0.02272554
#> 55  -0.50066890 0.02272554 0.02272554     -1  0.551869633 0.55186963 0.55186963
#> 56   0.13732280 0.13732280 0.13732280      0  0.815006441 0.81500644 0.81500644
#> 57  -1.98869101 0.02272554 0.02272554     -1 -1.950191464 0.02272554 0.02272554
#> 58   2.50135326 2.50135326 2.50135326      0  2.144972075 2.14497207 2.14497207
#> 59  -0.70424883 0.02272554 0.02272554     -1  0.078595267 0.07859527 0.07859527
#> 60  -0.21743431 0.02272554 0.02272554     -1  0.586977307 0.58697731 0.58697731
#> 61   1.88044124 1.88044124 1.88044124      0 -0.019619585 0.02272554 0.02272554
#> 62  -1.04387546 0.02272554 0.02272554     -1 -0.108091176 0.02272554 0.02272554
#> 63   1.11923685 1.11923685 1.11923685      0  0.810185342 0.81018534 0.81018534
#> 64  -0.34448588 0.02272554 0.02272554     -1 -0.081419201 0.02272554 0.02272554
#> 65  -0.12450381 0.02272554 0.02272554     -1 -1.915095668 0.02272554 0.02272554
#> 66  -0.45098200 0.02272554 0.02272554     -1 -1.239240846 0.02272554 0.02272554
#> 67   1.48887753 1.48887753 1.48887753      0  0.355855861 0.35585586 0.35585586
#> 68  -0.06779617 0.02272554 0.02272554     -1  0.295856399 0.29585640 0.29585640
#> 69   0.56610892 0.56610892 0.56610892      0  0.280221005 0.28022100 0.28022100
#> 70  -1.73060193 0.02272554 0.02272554     -1 -1.212932797 0.02272554 0.02272554
#> 71  -1.22224067 0.02272554 0.02272554     -1 -1.325149339 0.02272554 0.02272554
#> 72   0.73311621 0.73311621 0.73311621      0 -0.240953381 0.02272554 0.02272554
#> 73   0.52281730 0.52281730 0.52281730      0  1.793489601 1.79348960 1.79348960
#> 74  -0.27544740 0.02272554 0.02272554     -1  0.685417388 0.68541739 0.68541739
#> 75   0.75731488 0.75731488 0.75731488      0  1.526036245 1.52603625 1.52603625
#> 76  -0.06424620 0.02272554 0.02272554     -1  0.971684571 0.97168457 0.97168457
#> 77   0.76240276 0.76240276 0.76240276      0  0.288515685 0.28851568 0.28851568
#> 78   0.54246035 0.54246035 0.54246035      0 -0.732874522 0.02272554 0.02272554
#> 79   0.30192408 0.30192408 0.30192408      0 -0.003696597 0.02272554 0.02272554
#> 80  -1.18411047 0.02272554 0.02272554     -1  1.027659018 1.02765902 1.02765902
#> 81   2.18736441 2.18736441 2.18736441      0  1.145696029 1.14569603 1.14569603
#> 82   1.69770304 1.69770304 1.69770304      0  0.551179188 0.55117919 0.55117919
#> 83   1.48613591 1.48613591 1.48613591      0 -0.189191392 0.02272554 0.02272554
#> 84  -0.28126493 0.02272554 0.02272554     -1  1.244673727 1.24467373 1.24467373
#> 85  -0.75189748 0.02272554 0.02272554     -1 -0.197711961 0.02272554 0.02272554
#> 86  -0.03480370 0.02272554 0.02272554     -1  1.958306561 1.95830656 1.95830656
#> 87  -2.18528753 0.02272554 0.02272554     -1 -2.339408265 0.02272554 0.02272554
#> 88  -1.41703893 0.02272554 0.02272554     -1  1.147369404 1.14736940 1.14736940
#> 89  -0.15523999 0.02272554 0.02272554     -1  0.906759153 0.90675915 0.90675915
#> 90  -0.90212681 0.02272554 0.02272554     -1  0.240568073 0.24056807 0.24056807
#> 91  -0.19456221 0.02272554 0.02272554     -1  0.929276637 0.92927664 0.92927664
#> 92   1.11917408 1.11917408 1.11917408      0  0.722172589 0.72217259 0.72217259
#> 93   1.31984235 1.31984235 1.31984235      0  0.496581195 0.49658119 0.49658119
#> 94   0.25644424 0.25644424 0.25644424      0 -0.322440387 0.02272554 0.02272554
#> 95  -0.46087737 0.02272554 0.02272554     -1  1.302912003 1.30291200 1.30291200
#> 96   1.38226277 1.38226277 1.38226277      0  1.515254912 1.51525491 1.51525491
#> 97  -0.49513543 0.02272554 0.02272554     -1 -0.118636106 0.02272554 0.02272554
#> 98   0.25020748 0.25020748 0.25020748      0  1.388915137 1.38891514 1.38891514
#> 99   0.51977354 0.51977354 0.51977354      0  1.761036614 1.76103661 1.76103661
#> 100 -0.56294327 0.02272554 0.02272554     -1  0.049147675 0.04914768 0.04914768
#>     y_cens
#> 1       -1
#> 2       -1
#> 3       -1
#> 4       -1
#> 5       -1
#> 6        0
#> 7       -1
#> 8       -1
#> 9       -1
#> 10       0
#> 11      -1
#> 12      -1
#> 13       0
#> 14      -1
#> 15       0
#> 16      -1
#> 17      -1
#> 18      -1
#> 19       0
#> 20      -1
#> 21       0
#> 22       0
#> 23      -1
#> 24       0
#> 25       0
#> 26      -1
#> 27      -1
#> 28      -1
#> 29      -1
#> 30      -1
#> 31       0
#> 32       0
#> 33       0
#> 34       0
#> 35      -1
#> 36       0
#> 37      -1
#> 38      -1
#> 39       0
#> 40       0
#> 41      -1
#> 42      -1
#> 43       0
#> 44       0
#> 45       0
#> 46       0
#> 47      -1
#> 48       0
#> 49      -1
#> 50      -1
#> 51       0
#> 52      -1
#> 53      -1
#> 54      -1
#> 55       0
#> 56       0
#> 57      -1
#> 58       0
#> 59       0
#> 60       0
#> 61      -1
#> 62      -1
#> 63       0
#> 64      -1
#> 65      -1
#> 66      -1
#> 67       0
#> 68       0
#> 69       0
#> 70      -1
#> 71      -1
#> 72      -1
#> 73       0
#> 74       0
#> 75       0
#> 76       0
#> 77       0
#> 78      -1
#> 79      -1
#> 80       0
#> 81       0
#> 82       0
#> 83      -1
#> 84       0
#> 85      -1
#> 86       0
#> 87      -1
#> 88       0
#> 89       0
#> 90       0
#> 91       0
#> 92       0
#> 93       0
#> 94      -1
#> 95       0
#> 96       0
#> 97      -1
#> 98       0
#> 99       0
#> 100      0

# Generate 500 x and y values with a -0.25 correlation and 25% right and interval censored
generate_censored_data(500, rho = -0.25, L_prob = 0.25, direction = c(1, 2))

#>            x_true          x    x_upper x_cens        y_true          y
#> 1   -0.2074948745 -0.2074949 -0.2074949      0 -8.864044e-01 -0.8864044
#> 2   -0.8449469092 -0.8449469 -0.8449469      0  1.306131e+00  1.2155771
#> 3   -0.3150094382 -0.3150094 -0.3150094      0 -4.270449e-01 -0.4270449
#> 4    0.2181415267  0.2181415  0.2181415      0 -2.121712e-01 -0.2121712
#> 5    2.2531917428  1.2155771  1.2155771      1 -1.615349e+00 -1.6153491
#> 6    0.3395456395  0.3395456  0.3395456      0  4.430891e-01  0.4430891
#> 7    1.2578515817  1.2155771  1.2155771      1 -4.790435e-01 -0.4790435
#> 8    0.4353003741  0.4353004  0.4353004      0 -5.658422e-01 -0.5658422
#> 9    0.9107522471  0.9107522  0.9107522      0 -2.352172e+00 -2.3521719
#> 10  -2.3681346546 -2.3681347 -2.3681347      0  1.820195e+00  1.2155771
#> 11  -0.4800533587 -0.4800534 -0.4800534      0 -2.303736e+00 -2.3037363
#> 12   0.1859795914  0.1859796  0.1859796      0 -3.526162e-01 -0.3526162
#> 13   0.6024822613  0.6024823  0.6024823      0  1.845831e+00  1.2155771
#> 14   1.2619319484  1.2155771  1.2155771      1 -7.469257e-01 -0.7469257
#> 15  -0.1968742693 -0.1968743 -0.1968743      0  6.603268e-01  0.6603268
#> 16  -0.8069452583 -0.8069453 -0.8069453      0 -6.881924e-01 -0.6881924
#> 17   0.5021400535  0.5021401  0.5021401      0  1.385229e+00  1.2155771
#> 18  -0.1383822285 -0.1690979  0.1807250      2  5.184330e-01  0.5184330
#> 19   1.1604574765  1.1604575  1.1604575      0 -1.135830e+00 -1.1358301
#> 20   0.7196040288  0.7196040  0.7196040      0  2.207782e+00  1.2155771
#> 21  -0.2769766835 -0.2769767 -0.2769767      0 -8.128814e-01 -0.8128814
#> 22   0.9677806959  0.9677807  0.9677807      0  8.482873e-01  0.8482873
#> 23  -0.3047495892 -0.3047496 -0.3047496      0 -2.411809e-01 -0.2411809
#> 24  -0.9370815983 -0.9370816 -0.9370816      0 -7.405444e-01 -0.7405444
#> 25   0.8016450636  0.8016451  0.8016451      0 -7.770154e-01 -0.7770154
#> 26  -0.9081986221 -0.9081986 -0.9081986      0  1.041962e+00  1.0419623
#> 27  -0.4312007453 -0.4312007 -0.4312007      0  1.070576e-01 -0.1690979
#> 28   0.1876358234  0.1876358  0.1876358      0 -5.597724e-01 -0.5597724
#> 29  -0.4480164408 -0.4480164 -0.4480164      0  6.726205e-01  0.6726205
#> 30   0.8660339638  0.8660340  0.8660340      0 -1.551462e+00 -1.5514620
#> 31  -0.0734607856 -0.1690979  0.1807250      2  3.019416e-01  0.3019416
#> 32   1.5352858387  1.2155771  1.2155771      1 -5.404557e-01 -0.5404557
#> 33   0.4140690416  0.4140690  0.4140690      0  1.595225e+00  1.2155771
#> 34   0.9242415600  0.9242416  0.9242416      0 -1.546317e+00 -1.5463170
#> 35  -1.1304802107 -1.1304802 -1.1304802      0 -1.018135e+00 -1.0181350
#> 36   0.1428270746 -0.1690979  0.1807250      2 -8.683662e-01 -0.8683662
#> 37   0.0455787079 -0.1690979  0.1807250      2 -1.290793e+00 -1.2907931
#> 38  -0.6083049003 -0.6083049 -0.6083049      0  4.593507e-01  0.4593507
#> 39  -0.2516449500 -0.2516450 -0.2516450      0  1.572002e+00  1.2155771
#> 40   2.0335871928  1.2155771  1.2155771      1 -6.330109e-01 -0.6330109
#> 41   1.0380770750  1.0380771  1.0380771      0 -3.893987e-01 -0.3893987
#> 42  -0.6244879999 -0.6244880 -0.6244880      0  1.458874e+00  1.2155771
#> 43  -0.7458030581 -0.7458031 -0.7458031      0  9.937497e-01  0.9937497
#> 44  -1.9832007839 -1.9832008 -1.9832008      0 -7.987671e-02 -0.1690979
#> 45   1.7645731357  1.2155771  1.2155771      1 -4.984647e-01 -0.4984647
#> 46  -0.9977712147 -0.9977712 -0.9977712      0  1.188803e+00  1.1888028
#> 47   0.0259431984 -0.1690979  0.1807250      2  3.088574e-02 -0.1690979
#> 48   0.6455508872  0.6455509  0.6455509      0  5.224010e-01  0.5224010
#> 49  -0.6038862931 -0.6038863 -0.6038863      0  9.407302e-02 -0.1690979
#> 50  -0.1717510881 -0.1717511 -0.1717511      0  3.209391e-02 -0.1690979
#> 51   0.6171999197  0.6171999  0.6171999      0 -6.954841e-01 -0.6954841
#> 52   0.9610988740  0.9610989  0.9610989      0  1.648486e-01 -0.1690979
#> 53   0.3917033241  0.3917033  0.3917033      0 -1.495609e+00 -1.4956091
#> 54   1.0827962882  1.0827963  1.0827963      0 -1.057826e-01 -0.1690979
#> 55  -1.5601439325 -1.5601439 -1.5601439      0  7.417434e-01  0.7417434
#> 56   0.5600196539  0.5600197  0.5600197      0 -7.501176e-01 -0.7501176
#> 57  -0.3078918658 -0.3078919 -0.3078919      0  1.502818e-01 -0.1690979
#> 58   1.3183226279  1.2155771  1.2155771      1  5.592914e-01  0.5592914
#> 59   0.3542394903  0.3542395  0.3542395      0 -6.020797e-01 -0.6020797
#> 60  -0.7149175589 -0.7149176 -0.7149176      0  1.593719e+00  1.2155771
#> 61  -0.3398238214 -0.3398238 -0.3398238      0 -1.031600e-01 -0.1690979
#> 62   0.2764281571  0.2764282  0.2764282      0 -1.989851e+00 -1.9898515
#> 63  -0.2200850828 -0.2200851 -0.2200851      0 -2.363761e-01 -0.2363761
#> 64   1.1265447278  1.1265447  1.1265447      0  7.909701e-01  0.7909701
#> 65   0.7481811354  0.7481811  0.7481811      0 -6.848685e-01 -0.6848685
#> 66   1.3470877460  1.2155771  1.2155771      1 -1.976707e+00 -1.9767072
#> 67  -1.8335761845 -1.8335762 -1.8335762      0  1.160090e+00  1.1600896
#> 68   0.2339435199  0.2339435  0.2339435      0 -1.296798e+00 -1.2967978
#> 69  -0.3379684271 -0.3379684 -0.3379684      0 -5.001981e-01 -0.5001981
#> 70  -0.0900839648 -0.1690979  0.1807250      2  2.893254e-01  0.2893254
#> 71   1.0445509699  1.0445510  1.0445510      0  1.140841e+00  1.1408406
#> 72   2.3566443633  1.2155771  1.2155771      1 -1.086406e+00 -1.0864057
#> 73  -0.4151829588 -0.4151830 -0.4151830      0 -6.015362e-01 -0.6015362
#> 74  -0.4606993523 -0.4606994 -0.4606994      0 -2.831457e-01 -0.2831457
#> 75   0.0685889729 -0.1690979  0.1807250      2  8.105815e-02 -0.1690979
#> 76  -2.0559069886 -2.0559070 -2.0559070      0  9.130668e-01  0.9130668
#> 77  -1.1179875641 -1.1179876 -1.1179876      0  2.295306e+00  1.2155771
#> 78  -0.7172836007 -0.7172836 -0.7172836      0  4.048736e-01  0.4048736
#> 79  -0.2723647114 -0.2723647 -0.2723647      0  9.403425e-01  0.9403425
#> 80  -0.3260647472 -0.3260647 -0.3260647      0 -8.133898e-01 -0.8133898
#> 81  -0.4708153279 -0.4708153 -0.4708153      0  1.129300e+00  1.1293001
#> 82   1.0036273365  1.0036273  1.0036273      0 -4.495208e-01 -0.4495208
#> 83  -0.1763813754 -0.1763814 -0.1763814      0  7.144007e-01  0.7144007
#> 84   0.1224130491 -0.1690979  0.1807250      2  6.323030e-01  0.6323030
#> 85  -0.5603263630 -0.5603264 -0.5603264      0  1.943671e-02 -0.1690979
#> 86  -1.1887424940 -1.1887425 -1.1887425      0  5.976837e-01  0.5976837
#> 87   0.4123253071  0.4123253  0.4123253      0 -1.076314e+00 -1.0763139
#> 88   1.2305975999  1.2155771  1.2155771      1  1.575029e+00  1.2155771
#> 89  -1.3090325889 -1.3090326 -1.3090326      0  9.289307e-01  0.9289307
#> 90  -0.5564516528 -0.5564517 -0.5564517      0 -1.163187e+00 -1.1631869
#> 91  -0.7056815203 -0.7056815 -0.7056815      0 -9.809362e-01 -0.9809362
#> 92  -0.4740526349 -0.4740526 -0.4740526      0 -8.246615e-01 -0.8246615
#> 93   0.4109772783  0.4109773  0.4109773      0 -1.185228e+00 -1.1852278
#> 94  -1.0444267161 -1.0444267 -1.0444267      0 -2.843333e-01 -0.2843333
#> 95  -2.0990537339 -2.0990537 -2.0990537      0  4.377370e-01  0.4377370
#> 96   0.9757935633  0.9757936  0.9757936      0 -1.419671e+00 -1.4196705
#> 97  -0.2089815560 -0.2089816 -0.2089816      0 -2.447912e+00 -2.4479120
#> 98  -0.7783530289 -0.7783530 -0.7783530      0  6.079536e-01  0.6079536
#> 99  -1.3362444231 -1.3362444 -1.3362444      0 -3.494044e-01 -0.3494044
#> 100 -1.3689513939 -1.3689514 -1.3689514      0  1.970852e+00  1.2155771
#> 101 -0.6583990331 -0.6583990 -0.6583990      0 -1.221481e+00 -1.2214812
#> 102  1.0964014696  1.0964015  1.0964015      0 -5.866642e-01 -0.5866642
#> 103 -0.5327687199 -0.5327687 -0.5327687      0  1.170302e+00  1.1703021
#> 104 -1.4549916193 -1.4549916 -1.4549916      0  4.132434e-01  0.4132434
#> 105 -0.3616792043 -0.3616792 -0.3616792      0 -4.799376e-02 -0.1690979
#> 106  0.8244203100  0.8244203  0.8244203      0  1.913240e-01  0.1913240
#> 107 -0.3531893617 -0.3531894 -0.3531894      0  6.210361e-01  0.6210361
#> 108 -0.6145412656 -0.6145413 -0.6145413      0  2.506875e+00  1.2155771
#> 109 -1.6327798816 -1.6327799 -1.6327799      0  1.347133e+00  1.2155771
#> 110  1.2368886239  1.2155771  1.2155771      1 -1.274941e+00 -1.2749410
#> 111  0.2076996832  0.2076997  0.2076997      0 -6.459938e-01 -0.6459938
#> 112 -0.0605366684 -0.1690979  0.1807250      2 -1.909618e+00 -1.9096176
#> 113 -1.3176154318 -1.3176154 -1.3176154      0 -4.335912e-01 -0.4335912
#> 114  2.0190573449  1.2155771  1.2155771      1 -8.575832e-01 -0.8575832
#> 115 -0.9329557926 -0.9329558 -0.9329558      0 -3.110244e-01 -0.3110244
#> 116 -0.0564175681 -0.1690979  0.1807250      2  1.020230e-02 -0.1690979
#> 117  0.1355177113 -0.1690979  0.1807250      2  2.000731e+00  1.2155771
#> 118  0.7971604868  0.7971605  0.7971605      0  3.909324e-01  0.3909324
#> 119  0.9753570533  0.9753571  0.9753571      0 -6.948570e-01 -0.6948570
#> 120  0.4398214439  0.4398214  0.4398214      0  7.541158e-01  0.7541158
#> 121 -2.1630265862 -2.1630266 -2.1630266      0 -1.134443e+00 -1.1344427
#> 122 -1.1245509842 -1.1245510 -1.1245510      0 -5.807763e-01 -0.5807763
#> 123 -2.0009650013 -2.0009650 -2.0009650      0  3.352187e-01  0.3352187
#> 124  0.1668292951 -0.1690979  0.1807250      2  2.807119e-01  0.2807119
#> 125 -1.1489888235 -1.1489888 -1.1489888      0  2.213215e+00  1.2155771
#>        y_upper y_cens
#> 1   -0.8864044      0
#> 2    1.2155771      1
#> 3   -0.4270449      0
#> 4   -0.2121712      0
#> 5   -1.6153491      0
#> 6    0.4430891      0
#> 7   -0.4790435      0
#> 8   -0.5658422      0
#> 9   -2.3521719      0
#> 10   1.2155771      1
#> 11  -2.3037363      0
#> 12  -0.3526162      0
#> 13   1.2155771      1
#> 14  -0.7469257      0
#> 15   0.6603268      0
#> 16  -0.6881924      0
#> 17   1.2155771      1
#> 18   0.5184330      0
#> 19  -1.1358301      0
#> 20   1.2155771      1
#> 21  -0.8128814      0
#> 22   0.8482873      0
#> 23  -0.2411809      0
#> 24  -0.7405444      0
#> 25  -0.7770154      0
#> 26   1.0419623      0
#> 27   0.1807250      2
#> 28  -0.5597724      0
#> 29   0.6726205      0
#> 30  -1.5514620      0
#> 31   0.3019416      0
#> 32  -0.5404557      0
#> 33   1.2155771      1
#> 34  -1.5463170      0
#> 35  -1.0181350      0
#> 36  -0.8683662      0
#> 37  -1.2907931      0
#> 38   0.4593507      0
#> 39   1.2155771      1
#> 40  -0.6330109      0
#> 41  -0.3893987      0
#> 42   1.2155771      1
#> 43   0.9937497      0
#> 44   0.1807250      2
#> 45  -0.4984647      0
#> 46   1.1888028      0
#> 47   0.1807250      2
#> 48   0.5224010      0
#> 49   0.1807250      2
#> 50   0.1807250      2
#> 51  -0.6954841      0
#> 52   0.1807250      2
#> 53  -1.4956091      0
#> 54   0.1807250      2
#> 55   0.7417434      0
#> 56  -0.7501176      0
#> 57   0.1807250      2
#> 58   0.5592914      0
#> 59  -0.6020797      0
#> 60   1.2155771      1
#> 61   0.1807250      2
#> 62  -1.9898515      0
#> 63  -0.2363761      0
#> 64   0.7909701      0
#> 65  -0.6848685      0
#> 66  -1.9767072      0
#> 67   1.1600896      0
#> 68  -1.2967978      0
#> 69  -0.5001981      0
#> 70   0.2893254      0
#> 71   1.1408406      0
#> 72  -1.0864057      0
#> 73  -0.6015362      0
#> 74  -0.2831457      0
#> 75   0.1807250      2
#> 76   0.9130668      0
#> 77   1.2155771      1
#> 78   0.4048736      0
#> 79   0.9403425      0
#> 80  -0.8133898      0
#> 81   1.1293001      0
#> 82  -0.4495208      0
#> 83   0.7144007      0
#> 84   0.6323030      0
#> 85   0.1807250      2
#> 86   0.5976837      0
#> 87  -1.0763139      0
#> 88   1.2155771      1
#> 89   0.9289307      0
#> 90  -1.1631869      0
#> 91  -0.9809362      0
#> 92  -0.8246615      0
#> 93  -1.1852278      0
#> 94  -0.2843333      0
#> 95   0.4377370      0
#> 96  -1.4196705      0
#> 97  -2.4479120      0
#> 98   0.6079536      0
#> 99  -0.3494044      0
#> 100  1.2155771      1
#> 101 -1.2214812      0
#> 102 -0.5866642      0
#> 103  1.1703021      0
#> 104  0.4132434      0
#> 105  0.1807250      2
#> 106  0.1913240      0
#> 107  0.6210361      0
#> 108  1.2155771      1
#> 109  1.2155771      1
#> 110 -1.2749410      0
#> 111 -0.6459938      0
#> 112 -1.9096176      0
#> 113 -0.4335912      0
#> 114 -0.8575832      0
#> 115 -0.3110244      0
#> 116  0.1807250      2
#> 117  1.2155771      1
#> 118  0.3909324      0
#> 119 -0.6948570      0
#> 120  0.7541158      0
#> 121 -1.1344427      0
#> 122 -0.5807763      0
#> 123  0.3352187      0
#> 124  0.2807119      0
#> 125  1.2155771      1
#>  [ reached getOption("max.print") -- omitted 375 rows ]